# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Workforce API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/credentials": {
      "get": {
        "summary": "Get all credentials",
        "responses": {
          "200": {
            "description": "A list of credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/credential"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new credential",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/credential"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/credential"
                }
              }
            }
          }
        }
      }
    },
    "/credentials/{id}": {
      "get": {
        "summary": "Get a credential by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/credential"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a credential by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/credential"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/credential"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a credential by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The deleted credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/credential"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "credential": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The type of the credential",
            "enum": [
              "ai-worker",
              "api-resource",
              "coding-tool",
              "default-documentation",
              "discord-channel",
              "excalidraw-tool",
              "git-document-repository",
              "github-board-ticket-tool",
              "github-board-tracker",
              "github-pull-request-resource",
              "github-repo-resource",
              "google-drive-tool",
              "google-slides-tool",
              "human-worker",
              "internal-document-repository",
              "message-channel-tool",
              "native-channel",
              "openapi-channel-tool",
              "openapi-tool",
              "raw-text-resource",
              "simple-task",
              "slack-channel",
              "structured-task",
              "template-tool",
              "trello-ticket-tool",
              "trello-tracker",
              "web-service-tool",
              "worker-discord-token",
              "worker-native-token",
              "worker-slack-token"
            ]
          },
          "credential": {
            "type": "string"
          },
          "variables": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ai_worker_credential"
              },
              {
                "$ref": "#/components/schemas/api_resource_credential"
              },
              {
                "$ref": "#/components/schemas/coding_tool_credential"
              },
              {
                "$ref": "#/components/schemas/default_documentation_credential"
              },
              {
                "$ref": "#/components/schemas/discord_channel_credential"
              },
              {
                "$ref": "#/components/schemas/excalidraw_tool_credential"
              },
              {
                "$ref": "#/components/schemas/git_document_repository_credential"
              },
              {
                "$ref": "#/components/schemas/github_board_ticket_tool_credential"
              },
              {
                "$ref": "#/components/schemas/github_board_tracker_credential"
              },
              {
                "$ref": "#/components/schemas/github_pull_request_resource_credential"
              },
              {
                "$ref": "#/components/schemas/github_repo_resource_credential"
              },
              {
                "$ref": "#/components/schemas/google_drive_tool_credential"
              },
              {
                "$ref": "#/components/schemas/google_slides_tool_credential"
              },
              {
                "$ref": "#/components/schemas/human_worker_credential"
              },
              {
                "$ref": "#/components/schemas/internal_document_repository_credential"
              },
              {
                "$ref": "#/components/schemas/message_channel_tool_credential"
              },
              {
                "$ref": "#/components/schemas/native_channel_credential"
              },
              {
                "$ref": "#/components/schemas/openapi_channel_tool_credential"
              },
              {
                "$ref": "#/components/schemas/openapi_tool_credential"
              },
              {
                "$ref": "#/components/schemas/raw_text_resource_credential"
              },
              {
                "$ref": "#/components/schemas/simple_task_credential"
              },
              {
                "$ref": "#/components/schemas/slack_channel_credential"
              },
              {
                "$ref": "#/components/schemas/structured_task_credential"
              },
              {
                "$ref": "#/components/schemas/template_tool_credential"
              },
              {
                "$ref": "#/components/schemas/trello_ticket_tool_credential"
              },
              {
                "$ref": "#/components/schemas/trello_tracker_credential"
              },
              {
                "$ref": "#/components/schemas/web_service_tool_credential"
              },
              {
                "$ref": "#/components/schemas/worker_discord_token_credential"
              },
              {
                "$ref": "#/components/schemas/worker_native_token_credential"
              },
              {
                "$ref": "#/components/schemas/worker_slack_token_credential"
              }
            ]
          }
        },
        "required": [
          "type",
          "credential",
          "variables"
        ]
      },
...
      "discord_channel_credential": {
        "type": "object",
        "properties": {
          "bot_token": {
            "type": "string",
            "description": "The Discord bot token",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "bot_token"
        ],
        "description": "The sensitive variables for the discord-channel channel"
      },
      "native_channel_credential": {
        "type": "object",
        "properties": {
          "openai_token": {
            "type": "string",
            "description": "OpenAI API token",
            "x-workforce-sensitive": true
          }
        },
        "required": [],
        "description": "The sensitive variables for the native-channel channel"
      },
      "slack_channel_credential": {
        "type": "object",
        "properties": {
          "app_token": {
            "type": "string",
            "description": "The Slack app token",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "app_token"
        ],
        "description": "The sensitive variables for the slack-channel channel"
      },
      "worker_discord_token_credential": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "channel token",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "token"
        ],
        "description": "The sensitive variables for the worker-discord-token channel_user_credential"
      },
      "worker_native_token_credential": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "channel token",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "token"
        ],
        "description": "The sensitive variables for the worker-native-token channel_user_credential"
      },
      "worker_slack_token_credential": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "channel token",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "token"
        ],
        "description": "The sensitive variables for the worker-slack-token channel_user_credential"
      },
      "default_documentation_credential": {
        "type": "object",
        "properties": {},
        "required": [],
        "description": "The sensitive variables for the default-documentation documentation"
      },
      "git_document_repository_credential": {
        "type": "object",
        "properties": {
          "model_api_key": {
            "type": "string",
            "description": "The API key for the model.",
            "x-workforce-sensitive": true
          },
          "access_token": {
            "type": "string",
            "description": "The token to use for authentication.",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "model_api_key"
        ],
        "description": "The sensitive variables for the git-document-repository document_repository"
      },
      "internal_document_repository_credential": {
        "type": "object",
        "properties": {
          "model_api_key": {
            "type": "string",
            "description": "The API key for the model.",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "model_api_key"
        ],
        "description": "The sensitive variables for the internal-document-repository document_repository"
      },
      "api_resource_credential": {
        "type": "object",
        "properties": {
          "custom_headers": {
            "type": "string",
            "description": "Custom headers to include in the request. Provide a JSON object.",
            "x-workforce-multiline": true,
            "x-workforce-sensitive": true
          },
          "username": {
            "type": "string",
            "description": "The username to use for basic authentication.",
            "x-workforce-sensitive": true
          },
          "password": {
            "type": "string",
            "description": "The password to use for basic authentication.",
            "x-workforce-sensitive": true
          },
          "bearer_token": {
            "type": "string",
            "description": "The bearer token to use for bearer token authentication.",
            "x-workforce-sensitive": true
          },
          "api_key": {
            "type": "string",
            "description": "The API key to use for API key authentication.",
            "x-workforce-sensitive": true
          },
          "mtls_cert": {
            "type": "string",
            "description": "The certificate to use for mTLS authentication.",
            "x-workforce-multiline": true,
            "x-workforce-sensitive": true
          },
          "mtls_key": {
            "type": "string",
            "description": "The key to use for mTLS authentication.",
            "x-workforce-multiline": true,
            "x-workforce-sensitive": true
          },
          "mtls_ca": {
            "type": "string",
            "description": "The CA to use for mTLS authentication.",
            "x-workforce-multiline": true,
            "x-workforce-sensitive": true
          },
          "client_id": {
            "type": "string",
            "description": "The client ID to use for Oauth2 authentication.",
            "x-workforce-sensitive": true
          },
          "client_secret": {
            "type": "string",
            "description": "The client secret to use for Oauth2 authentication.",
            "x-workforce-sensitive": true
          },
          "scopes": {
            "type": "string",
            "description": "The scope to use for Oauth2 authentication. Provide a comma-separated list of scopes.",
            "x-workforce-sensitive": true
          },
          "oauth2_audience": {
            "type": "string",
            "description": "The audience to use for Oauth2 authentication.",
            "x-workforce-sensitive": true
          }
        },
        "required": [],
        "description": "The sensitive variables for the api-resource resource"
      },
      "github_pull_request_resource_credential": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The username of the authenticated user.",
            "x-workforce-sensitive": true
          },
          "access_token": {
            "type": "string",
            "description": "The token to use for authentication.",
            "x-workforce-sensitive": true
          }
        },
        "required": [],
        "description": "The sensitive variables for the github-pull-request-resource resource"
      },
      "github_repo_resource_credential": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "The token to use for authentication.",
            "x-workforce-sensitive": true
          }
        },
        "required": [],
        "description": "The sensitive variables for the github-repo-resource resource"
      },
      "raw_text_resource_credential": {
        "type": "object",
        "properties": {},
        "required": [],
        "description": "The sensitive variables for the raw-text-resource resource"
      },
      "simple_task_credential": {
        "type": "object",
        "properties": {},
        "required": [],
        "description": "The sensitive variables for the simple-task task"
      },
      "structured_task_credential": {
        "type": "object",
        "properties": {},
        "required": [],
        "description": "The sensitive variables for the structured-task task"
      },
      "coding_tool_credential": {
        "type": "object",
        "properties": {
          "index_repo_username": {
            "type": "string",
            "description": "The username to use when cloning the mono repository.",
            "x-workforce-sensitive": true
          },
          "index_repo_password": {
            "type": "string",
            "description": "The password to use when cloning the mono repository.",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "index_repo_username",
          "index_repo_password"
        ],
        "description": "The sensitive variables for the coding-tool tool"
      },
      "excalidraw_tool_credential": {
        "type": "object",
        "properties": {},
        "required": [],
        "description": "The sensitive variables for the excalidraw-tool tool"
      },
      "github_board_ticket_tool_credential": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "The Github access token",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "access_token"
        ],
        "description": "The sensitive variables for the github-board-ticket-tool tool"
      },
      "google_drive_tool_credential": {
        "type": "object",
        "properties": {
          "client_id": {
            "type": "string",
            "description": "The client ID for the Google Drive API.",
            "x-workforce-sensitive": true
          },
          "client_secret": {
            "type": "string",
            "description": "The client secret for the Google Drive API.",
            "x-workforce-sensitive": true
          },
          "oauth2_audience": {
            "type": "string",
            "description": "The audience for the OAuth2 token.",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "client_id",
          "client_secret"
        ],
        "description": "The sensitive variables for the google-drive-tool tool"
      },
      "google_slides_tool_credential": {
        "type": "object",
        "properties": {
          "client_id": {
            "type": "string",
            "description": "The client ID for the Google Slides API.",
            "x-workforce-sensitive": true
          },
          "client_secret": {
            "type": "string",
            "description": "The client secret for the Google Slides API.",
            "x-workforce-sensitive": true
          },
          "oauth2_audience": {
            "type": "string",
            "description": "The audience for the OAuth2 token.",
            "x-workforce-sensitive": true
          }
        },
        "required": [],
        "description": "The sensitive variables for the google-slides-tool tool"
      },
      "message_channel_tool_credential": {
        "type": "object",
        "properties": {},
        "required": [],
        "description": "The sensitive variables for the message-channel-tool tool"
      },
      "openapi_channel_tool_credential": {
        "type": "object",
        "properties": {},
        "required": [],
        "description": "The sensitive variables for the openapi-channel-tool tool"
      },
      "openapi_tool_credential": {
        "type": "object",
        "properties": {
          "custom_headers": {
            "type": "string",
            "description": "Custom headers to include in the request. Provide a JSON object.",
            "x-workforce-multiline": true,
            "x-workforce-sensitive": true
          },
          "username": {
            "type": "string",
            "description": "The username to use for basic authentication.",
            "x-workforce-sensitive": true
          },
          "password": {
            "type": "string",
            "description": "The password to use for basic authentication.",
            "x-workforce-sensitive": true
          },
          "bearer_token": {
            "type": "string",
            "description": "The bearer token to use for bearer token authentication.",
            "x-workforce-sensitive": true
          },
          "api_key": {
            "type": "string",
            "description": "The API key to use for API key authentication.",
            "x-workforce-sensitive": true
          },
          "mtls_cert": {
            "type": "string",
            "description": "The certificate to use for mTLS authentication.",
            "x-workforce-multiline": true,
            "x-workforce-sensitive": true
          },
          "mtls_key": {
            "type": "string",
            "description": "The key to use for mTLS authentication.",
            "x-workforce-multiline": true,
            "x-workforce-sensitive": true
          },
          "mtls_ca": {
            "type": "string",
            "description": "The CA to use for mTLS authentication.",
            "x-workforce-multiline": true,
            "x-workforce-sensitive": true
          },
          "client_id": {
            "type": "string",
            "description": "The client ID to use for Oauth2 authentication.",
            "x-workforce-sensitive": true
          },
          "client_secret": {
            "type": "string",
            "description": "The client secret to use for Oauth2 authentication.",
            "x-workforce-sensitive": true
          },
          "scopes": {
            "type": "string",
            "description": "The scope to use for Oauth2 authentication. Provide a comma-separated list of scopes.",
            "x-workforce-sensitive": true
          },
          "oauth2_audience": {
            "type": "string",
            "description": "The audience to use for Oauth2 authentication.",
            "x-workforce-sensitive": true
          }
        },
        "required": [],
        "description": "The sensitive variables for the openapi-tool tool"
      },
      "template_tool_credential": {
        "type": "object",
        "properties": {},
        "required": [],
        "description": "The sensitive variables for the template-tool tool"
      },
      "trello_ticket_tool_credential": {
        "type": "object",
        "properties": {
          "api_key": {
            "type": "string",
            "description": "The Trello API key.",
            "x-workforce-sensitive": true
          },
          "api_token": {
            "type": "string",
            "description": "The Trello API token.",
            "x-workforce-sensitive": true
          },
          "app_secret": {
            "type": "string",
            "description": "The Trello app secret.",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "api_key",
          "api_token",
          "app_secret"
        ],
        "description": "The sensitive variables for the trello-ticket-tool tool"
      },
      "web_service_tool_credential": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "The password to use for basic authentication.",
            "x-workforce-sensitive": true
          },
          "bearer_token": {
            "type": "string",
            "description": "Bearer token to use for Oauth2 authentication.",
            "x-workforce-sensitive": true
          },
          "client_id": {
            "type": "string",
            "description": "The client ID to use for Oauth2 Client Credentials authentication.",
            "x-workforce-sensitive": true
          },
          "client_secret": {
            "type": "string",
            "description": "The client secret to use for Oauth2 Client Credentials authentication.",
            "x-workforce-sensitive": true
          },
          "scope": {
            "type": "string",
            "description": "The scope to use for Oauth2 Client Credentials authentication.",
            "x-workforce-sensitive": true
          }
        },
        "required": [],
        "description": "The sensitive variables for the web-service-tool tool"
      },
      "github_board_tracker_credential": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "The Github access token",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "access_token"
        ],
        "description": "The sensitive variables for the github-board-tracker tracker"
      },
      "trello_tracker_credential": {
        "type": "object",
        "properties": {
          "api_key": {
            "type": "string",
            "description": "The Trello API key",
            "x-workforce-sensitive": true
          },
          "api_token": {
            "type": "string",
            "description": "The Trello API token",
            "x-workforce-sensitive": true
          },
          "app_secret": {
            "type": "string",
            "description": "The secret used to verify the webhook",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "api_key",
          "api_token",
          "app_secret"
        ],
        "description": "The sensitive variables for the trello-tracker tracker"
      },
      "ai_worker_credential": {
        "type": "object",
        "properties": {
          "api_token": {
            "type": "string",
            "description": "The API token",
            "x-workforce-sensitive": true
          }
        },
        "required": [
          "api_token"
        ],
        "description": "The sensitive variables for the ai-worker worker"
      },
      "human_worker_credential": {
        "type": "object",
        "properties": {},
        "required": [],
        "description": "The sensitive variables for the human-worker worker"
      }
    }
  }
}